backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# HTTP Client
httpx==0.27.2

# Environment
python-dotenv==1.0.1

# Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Utils
python-dateutil==2.9.0
